api_server/requirements.txt
dj-rest-auth==2.2.3
Django==4.0.3
djangorestframework==3.13.1
ulid-py==1.1.0